Refactorised Java Folder
