+++ Java/Array/NextGreaterElement2
