+++ Java/PatternQuestions/Question15.java
