+++ Java/String/Optimal Partition of String.java
